backend-api/azure_functions/requirements.txt
azure-functions
azure-cosmos
azure-identity
bcrypt 
jwt
